findKthMax
In this problem, you have to implement the findKthMax() function to take a BST and any number “k” as an input, and return kth maximum number from that tree. An illustration is also provided for your understanding.

Function Prototype #
int findKthMax(Node root, int k);
Here, root is the root node of the Binary Search Tree and k is an integer number.

Output #
It returns the kth maximum value from the given tree.